+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.0] - 2021-01-11
+
+### Added 
+- `queueSize` option for `useAsyncEffect`;
+
+### Updated
+- reworked `cancelPrevious` and `combine` logic;
+
 ## [0.3.0] - 2021-01-07
 
 ### Fixed